+++ src/map.cpp
@@ -267,6 +267,8 @@ std::string to_string(MapManager::Type t)
       return "elapsed";
     case MapManager::Type::SeqPrintfData:
       return "seq_printf_data";
+    case MapManager::Type::DebugfData:
+      return "debugf_data";
   }
   return {}; // unreached
 }